Phi-3.5-vision-instruct-ChartCap/requirements.txt
tqdm
torch==2.3.0
torchvision==0.18.0
numpy==1.24.4
Pillow==10.3.0
transformers==4.43.0
accelerate==0.30.0